requirements.txt
cuda-python~=12.2.0
numpy~=1.24.0
pillow~=10.0.0
tqdm~=4.65.0
joblib~=1.3.1